:truck: modify : elements폴더 파일명, 경로 정리
elements폴더 내에 있는 하위 폴더와 해당 파일들의 styled-components를 분리, 이에 맞도록 관련된 컴포넌트의 경로 수정
