+++ src/elements/image/style.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const EmojiImg = styled.img<{ width: number; height: number }>`
+  width: ${props => props.width}px;
+  height: ${props => props.height}px;
+`;
+
+export default EmojiImg;
